fix: update panels to include rename from PanelInner to AssetPanel
